+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/rest/service/RestPutDecoderAction.java
@@ -19,7 +19,6 @@ package com.wazuh.contentmanager.rest.service;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import org.opensearch.core.rest.RestStatus;
-import org.opensearch.rest.NamedRoute;
 import org.opensearch.transport.client.Client;
 
 import java.util.List;
@@ -58,7 +57,6 @@ import static org.opensearch.rest.RestRequest.Method.PUT;
 public class RestPutDecoderAction extends AbstractUpdateAction {
 
     private static final String ENDPOINT_NAME = "content_manager_decoder_update";
-    private static final String ENDPOINT_UNIQUE_NAME = "plugin:content_manager/decoder_update";
 
     public RestPutDecoderAction(EngineService engine) {
         super(engine);
@@ -77,12 +75,7 @@ public class RestPutDecoderAction extends AbstractUpdateAction {
      */
     @Override
     public List<Route> routes() {
-        return List.of(
-                new NamedRoute.Builder()
-                        .path(PluginSettings.DECODERS_URI + "/{id}")
-                        .method(PUT)
-                        .uniqueName(ENDPOINT_UNIQUE_NAME)
-                        .build());
+        return List.of(new Route(PUT, PluginSettings.DECODERS_URI + "/{id}"));
     }
 
     @Override